pull native gatt characteristic implementation out of native gatt service
