+++ buruberi-core/src/main/java/is/hello/buruberi/bluetooth/stacks/android/NativeGattService.java
@@ -16,7 +16,6 @@
 package is.hello.buruberi.bluetooth.stacks.android;
 
 import android.bluetooth.BluetoothGattCharacteristic;
-import android.bluetooth.BluetoothGattDescriptor;
 import android.bluetooth.BluetoothGattService;
 import android.support.annotation.NonNull;
 
@@ -27,18 +26,16 @@ import java.util.Map;
 import java.util.UUID;
 
 import is.hello.buruberi.bluetooth.stacks.GattCharacteristic;
-import is.hello.buruberi.bluetooth.stacks.GattPeripheral;
 import is.hello.buruberi.bluetooth.stacks.GattService;
-import is.hello.buruberi.bluetooth.stacks.OperationTimeout;
-import rx.Observable;
 
 public final class NativeGattService implements GattService {
-    final @NonNull BluetoothGattService service;
-    final @NonNull NativeGattPeripheral peripheral;
+    final BluetoothGattService service;
+    private final NativeGattPeripheral peripheral;
+    private final Map<UUID, NativeGattCharacteristic> characteristics = new HashMap<>();
 
 
     static @NonNull Map<UUID, GattService> wrap(@NonNull List<BluetoothGattService> services,
-                                                      @NonNull NativeGattPeripheral peripheral) {
+                                                @NonNull NativeGattPeripheral peripheral) {
         final Map<UUID, GattService> peripheralServices = new HashMap<>();
 
         for (final BluetoothGattService nativeService : services) {
@@ -79,12 +76,15 @@ public final class NativeGattService implements GattService {
 
     @Override
     public GattCharacteristic getCharacteristic(@NonNull UUID identifier) {
-        final BluetoothGattCharacteristic characteristic = service.getCharacteristic(identifier);
-        if (characteristic != null) {
-            return new NativeGattCharacteristic(characteristic);
-        } else {
-            return null;
+        NativeGattCharacteristic nativeCharacteristic = characteristics.get(identifier);
+        if (nativeCharacteristic == null) {
+            final BluetoothGattCharacteristic characteristic = service.getCharacteristic(identifier);
+            if (characteristic != null) {
+                nativeCharacteristic = new NativeGattCharacteristic(characteristic, this, peripheral);
+                characteristics.put(identifier, nativeCharacteristic);
+            }
         }
+        return nativeCharacteristic;
     }
 
 
@@ -93,8 +93,7 @@ public final class NativeGattService implements GattService {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
 
-        NativeGattService that = (NativeGattService) o;
-
+        final NativeGattService that = (NativeGattService) o;
         return service.equals(that.service);
     }
 
@@ -109,106 +108,5 @@ public final class NativeGattService implements GattService {
         return "AndroidPeripheralService{" +
                 "service=" + service +
                 '}';
-    }
-
-
-    class NativeGattCharacteristic implements GattCharacteristic {
-        private final BluetoothGattCharacteristic nativeCharacteristic;
-
-        NativeGattCharacteristic(@NonNull BluetoothGattCharacteristic nativeCharacteristic) {
-            this.nativeCharacteristic = nativeCharacteristic;
-        }
-
-        @Override
-        public UUID getUuid() {
-            return nativeCharacteristic.getUuid();
-        }
-
-        @Override
-        @Properties
-        public int getProperties() {
-            final @Properties int properties = nativeCharacteristic.getProperties();
-            return properties;
-        }
-
-        @Override
-        @Permissions
-        public int getPermissions() {
-            final @Permissions int permissions = nativeCharacteristic.getPermissions();
-            return permissions;
-        }
-
-        @NonNull
-        @Override
-        public GattService getService() {
-            return NativeGattService.this;
-        }
-
-        @NonNull
-        @Override
-        public List<UUID> getDescriptors() {
-            final List<UUID> identifiers = new ArrayList<>();
-            for (final BluetoothGattDescriptor descriptor : nativeCharacteristic.getDescriptors()) {
-                identifiers.add(descriptor.getUuid());
-            }
-            return identifiers;
-        }
-
-        @Permissions
-        @Override
-        public int getDescriptorPermissions(@NonNull UUID identifier) {
-            final BluetoothGattDescriptor descriptor = nativeCharacteristic.getDescriptor(identifier);
-            if (descriptor != null) {
-                final @Permissions int permissions = descriptor.getPermissions();
-                return permissions;
-            } else {
-                return PERMISSION_NULL;
-            }
-        }
-
-        @NonNull
-        @Override
-        public Observable<UUID> enableNotification(@NonNull UUID descriptor,
-                                                   @NonNull OperationTimeout timeout) {
-            // TODO: Re-implement this operation in terms of new API.
-
-            return peripheral.enableNotification(getService(),
-                                                 getUuid(),
-                                                 descriptor,
-                                                 timeout);
-        }
-
-        @NonNull
-        @Override
-        public Observable<UUID> disableNotification(@NonNull UUID descriptor,
-                                                    @NonNull OperationTimeout timeout) {
-            // TODO: Re-implement this operation in terms of new API.
-
-            return peripheral.disableNotification(getService(),
-                                                  getUuid(),
-                                                  descriptor,
-                                                  timeout);
-        }
-
-        @NonNull
-        @Override
-        public Observable<Void> write(@NonNull GattPeripheral.WriteType writeType,
-                                      @NonNull byte[] payload,
-                                      @NonNull OperationTimeout timeout) {
-            // TODO: Re-implement this operation in terms of new API.
-
-            return peripheral.writeCommand(getService(),
-                                           getUuid(),
-                                           writeType,
-                                           payload,
-                                           timeout);
-        }
-
-        @Override
-        public String toString() {
-            return "NativeCharacteristic{" +
-                    "uuid=" + getUuid() +
-                    '}';
-        }
     }
 }
